Separated tzbd as an independent target
